Refactor `TfrtCpuBuffer` by adding a new intermediate class `AbstractTfrtCpuBuffer` between `PjRtBuffer` and `TfrtCpuBuffer`. Other clients may use this new buffer class as a shared impl to create their own host buffer classes.
PiperOrigin-RevId: 528915214
