+++ tensorflow/compiler/xla/pjrt/tfrt_cpu_pjrt_client.cc
@@ -31,7 +31,6 @@ limitations under the License.
 #include "absl/algorithm/container.h"
 #include "absl/base/casts.h"
 #include "absl/base/dynamic_annotations.h"
-#include "absl/base/thread_annotations.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/inlined_vector.h"
 #include "absl/functional/any_invocable.h"
@@ -69,8 +68,6 @@ limitations under the License.
 #include "tensorflow/compiler/xla/primitive_util.h"
 #include "tensorflow/compiler/xla/runtime/cpu_event.h"
 #include "tensorflow/compiler/xla/service/buffer_assignment.h"
-#include "tensorflow/compiler/xla/service/compiler.h"
-#include "tensorflow/compiler/xla/service/computation_layout.h"
 #include "tensorflow/compiler/xla/service/computation_placer.h"
 #include "tensorflow/compiler/xla/service/cpu/cpu_compiler.h"
 #include "tensorflow/compiler/xla/service/cpu/cpu_executable.h"
@@ -80,18 +77,9 @@ limitations under the License.
 #include "tensorflow/compiler/xla/service/executable.h"
 #include "tensorflow/compiler/xla/service/hlo.pb.h"
 #include "tensorflow/compiler/xla/service/hlo_cost_analysis.h"
-#include "tensorflow/compiler/xla/service/hlo_dataflow_analysis.h"
-#include "tensorflow/compiler/xla/service/hlo_module_config.h"
 #include "tensorflow/compiler/xla/service/hlo_module_util.h"
-#include "tensorflow/compiler/xla/service/hlo_value.h"
-#include "tensorflow/compiler/xla/service/shaped_buffer.h"
 #include "tensorflow/compiler/xla/shape.h"
-#include "tensorflow/compiler/xla/shape_tree.h"
-#include "tensorflow/compiler/xla/shape_util.h"
-#include "tensorflow/compiler/xla/status.h"
-#include "tensorflow/compiler/xla/stream_executor/device_memory.h"
-#include "tensorflow/compiler/xla/util.h"
-#include "tensorflow/compiler/xla/xla.pb.h"
+#include "tensorflow/compiler/xla/statusor.h"
 #include "tensorflow/compiler/xla/xla_data.pb.h"
 #include "tensorflow/tsl/platform/casts.h"
 #include "tensorflow/tsl/platform/denormal.h"
@@ -101,45 +89,12 @@ limitations under the License.
 #include "tensorflow/tsl/platform/statusor.h"
 #include "tensorflow/tsl/platform/threadpool.h"
 #include "tensorflow/tsl/profiler/lib/connected_traceme.h"
-#include "tensorflow/tsl/profiler/lib/context_types.h"
-#include "tensorflow/tsl/profiler/lib/traceme.h"
-#include "tfrt/concurrency/async_value.h"  // from @tf_runtime
-#include "tfrt/concurrency/async_value_ref.h"  // from @tf_runtime
-#include "tfrt/concurrency/ref_count.h"  // from @tf_runtime
 #include "tfrt/host_context/async_value_ref.h"  // from @tf_runtime
 #include "tfrt/support/forward_decls.h"  // from @tf_runtime
 
 namespace xla {
-namespace {
 
 using ::xla::runtime::CpuEvent;
-
-// A RAII helper class used to set an AsyncValueRef<CpuEvent> to a ready state
-// upon destruction. In many cases in PjRt implementation, there will be
-// multiple return statements in the function, all of which require setting some
-// AsyncValueRef<CpuEvent> to be ready. This class could make such code more
-// robust by using setting the AsyncValue in the destructor.
-class MarkEventReadyOnExit {
- public:
-  explicit MarkEventReadyOnExit(tfrt::AsyncValueRef<CpuEvent> event)
-      : event_(std::move(event)) {}
-
-  MarkEventReadyOnExit(const MarkEventReadyOnExit&) = delete;
-  MarkEventReadyOnExit& operator=(const MarkEventReadyOnExit&) = delete;
-  MarkEventReadyOnExit(MarkEventReadyOnExit&&) = default;
-  MarkEventReadyOnExit& operator=(MarkEventReadyOnExit&&) = default;
-
-  ~MarkEventReadyOnExit() {
-    if (event_) event_.SetStateConcrete();
-  }
-
-  tfrt::AsyncValueRef<CpuEvent> Release() && { return std::move(event_); }
-
- private:
-  tfrt::AsyncValueRef<CpuEvent> event_;
-};
-
-}  // namespace
 
 static const char kCpuPlatformName[] = "cpu";
 static constexpr size_t kSmallDataTransferByteSize = 102400;  // 100 KiB
@@ -837,198 +792,10 @@ TfrtCpuBuffer::TfrtCpuBuffer(
     Shape on_device_shape,
     std::unique_ptr<TrackedTfrtCpuDeviceBuffer> tracked_device_buffer,
     TfrtCpuClient* client, TfrtCpuDevice* device)
-    : client_(client),
-      on_device_shape_(std::move(on_device_shape)),
-      device_(device),
-      tracked_device_buffer_(std::move(tracked_device_buffer)) {}
-
-TfrtCpuBuffer::~TfrtCpuBuffer() {
-  Delete();
-  CHECK_EQ(external_reference_counter_, 0);
-}
-
-StatusOr<size_t> TfrtCpuBuffer::GetOnDeviceSizeInBytes() const {
-  return ShapeUtil::ByteSizeOf(on_device_shape_);
-}
-
-StatusOr<std::unique_ptr<PjRtBuffer::ExternalReference>>
-TfrtCpuBuffer::AcquireExternalReference() {
-  class ScopedExternalReference : public PjRtBuffer::ExternalReference {
-   public:
-    explicit ScopedExternalReference(TfrtCpuBuffer* buffer,
-                                     std::shared_ptr<MaybeOwningCpuMemory> data)
-        : buffer_(buffer), data_(std::move(data)) {
-      DCHECK(data_);
-      data_ptr_ = data_->data();
-    }
-
-    ~ScopedExternalReference() override { buffer_->DropExternalReference(); }
-
-   private:
-    TfrtCpuBuffer* buffer_ = nullptr;
-    // Keep a reference to the underlying data used. Note that it is still
-    // users' responsibility to synchronize reads and writes to the data.
-    std::shared_ptr<MaybeOwningCpuMemory> data_;
-  };
-
-  absl::MutexLock lock(&mu_);
-  if (tracked_device_buffer_ == nullptr) {
-    return InvalidArgument("Buffer has been deleted or donated.");
-  }
-
-  ++external_reference_counter_;
-
-  return {std::make_unique<ScopedExternalReference>(
-      this, tracked_device_buffer_->Buffers()[0])};
-}
-
-class TrackedCpuDeviceBufferExternalReference
-    : public PjRtBuffer::ExternalReference {
- public:
-  explicit TrackedCpuDeviceBufferExternalReference(
-      std::unique_ptr<TrackedTfrtCpuDeviceBuffer> tracked_device_buffer)
-      : tracked_device_buffer_(std::move(tracked_device_buffer)) {
-    data_ptr_ = tracked_device_buffer_->Buffers()[0]->data();
-  }
-
-  ~TrackedCpuDeviceBufferExternalReference() override = default;
-
- private:
-  std::unique_ptr<TrackedTfrtCpuDeviceBuffer> tracked_device_buffer_;
-};
-
-StatusOr<std::unique_ptr<PjRtBuffer::ExternalReference>>
-TfrtCpuBuffer::ReleaseDeviceMemoryOwnership(
-    bool wait_for_operations_to_complete) {
-  if (on_device_shape_.IsTuple()) {
-    return InvalidArgument(
-        "ReleaseDeviceMemoryOwnership allowed only for non-tuple");
-  }
-  TF_ASSIGN_OR_RETURN(
-      std::unique_ptr<TrackedTfrtCpuDeviceBuffer> tracked_device_buffer,
-      Release(wait_for_operations_to_complete));
-
-  std::unique_ptr<PjRtBuffer::ExternalReference> ref;
-  if (tracked_device_buffer) {
-    ref = std::make_unique<TrackedCpuDeviceBufferExternalReference>(
-        std::move(tracked_device_buffer));
-  }
-  return ref;
-}
-
-void TfrtCpuBuffer::CommitDonation() {
-  absl::MutexLock lock(&mu_);
-  CHECK(pending_donation_);
-  CHECK(!tracked_device_buffer_);
-  pending_donation_ = false;
-}
-
-void TfrtCpuBuffer::AbortDonation(
-    std::unique_ptr<TrackedTfrtCpuDeviceBuffer> device_buffer) {
-  absl::MutexLock lock(&mu_);
-  CHECK(pending_donation_);
-  CHECK(!tracked_device_buffer_);
-  pending_donation_ = false;
-  tracked_device_buffer_ = std::move(device_buffer);
-}
-
-void TfrtCpuBuffer::Delete() {
-  auto device_buffer = ReleaseBufferLocked();
-  if (device_buffer == nullptr) return;
-
-  // Now that all holds have completed and no more can be added, we can get
-  // the final set of usage events.
-  absl::InlinedVector<tfrt::AsyncValueRef<CpuEvent>, 4> usage_events =
-      device_buffer->LockUseAndTransferUsageEvents();
-
-  std::vector<tfrt::AsyncValue*> event_avs;
-  event_avs.reserve(usage_events.size() + 1);
-  for (auto& event : usage_events) {
-    event_avs.push_back(event.GetAsyncValue());
-  }
-
-  // We should also wait for the definition event.
-  event_avs.push_back(device_buffer->definition_event().GetAsyncValue());
-
-  RunWhenReady(event_avs, [device_buffer = std::move(device_buffer)]() mutable {
-    device_buffer.reset();
-  });
-}
-
-bool TfrtCpuBuffer::IsDeleted() {
-  absl::MutexLock lock(&mu_);
-  return tracked_device_buffer_ == nullptr;
-}
-
-std::unique_ptr<TrackedTfrtCpuDeviceBuffer>
-TfrtCpuBuffer::ReleaseBufferLocked() {
-  absl::MutexLock lock(&mu_);
-  auto condition = [this]() ABSL_SHARED_LOCKS_REQUIRED(mu_) {
-    return !pending_donation_;
-  };
-  mu_.Await(absl::Condition(&condition));
-  return std::move(tracked_device_buffer_);
-}
-
-StatusOr<std::unique_ptr<TrackedTfrtCpuDeviceBuffer>> TfrtCpuBuffer::Release(
-    bool wait_for_operations_to_complete) {
-  std::unique_ptr<TrackedTfrtCpuDeviceBuffer> device_buffer =
-      ReleaseBufferLocked();
-  if (device_buffer == nullptr) return {nullptr};
-
-  absl::InlinedVector<tfrt::AsyncValueRef<CpuEvent>, 4> events;
-  // Now that all holds have completed and no more can be added, we can get
-  // the final set of usage events.
-  events = device_buffer->LockUseAndTransferUsageEvents();
-
-  if (wait_for_operations_to_complete) {
-    // Block the host until all usage events have completed. Usage events
-    // dominate definition events, so this also waits for the buffer to be
-    // defined. Return the first error encountered.
-    Status first_error;
-    for (const auto& av : events) {
-      BlockUntilReady(av.GetAsyncValue());
-      if (auto* error = av.GetErrorIfPresent()) {
-        first_error.Update(
-            InternalError("Error Execute: %s", error->message()));
-      }
-    }
-    if (!first_error.ok()) return std::move(first_error);
-  }
-
-  return device_buffer;
-}
-
-TrackedTfrtCpuDeviceBuffer* TfrtCpuBuffer::AcquireUsage(
-    tfrt::AsyncValueRef<CpuEvent> usage_event) {
-  absl::MutexLock lock(&mu_);
-  if (!tracked_device_buffer_) {
-    return nullptr;
-  }
-
-  tracked_device_buffer_->AddUsageEvents(absl::MakeSpan(&usage_event, 1));
-  return tracked_device_buffer_.get();
-}
-
-StatusOr<TfrtCpuBuffer::DonationTransaction> TfrtCpuBuffer::AcquireDonation() {
-  absl::MutexLock lock(&mu_);
-
-  if (tracked_device_buffer_ == nullptr) {
-    return InvalidArgument("Donation requested for invalid buffer");
-  }
-
-  if (external_reference_counter_ > 0) {
-    return InvalidArgument(
-        "Donation requested for buffer with external reference");
-  }
-
-  CHECK(!pending_donation_);
-  pending_donation_ = true;
-
-  // Swap out `tracked_device_buffer_` so that no one can acquire a usage event
-  // after this point.
-  return DonationTransaction(this, std::move(tracked_device_buffer_));
-}
+    : AbstractTfrtCpuBuffer(std::move(on_device_shape),
+                            std::move(tracked_device_buffer)),
+      client_(client),
+      device_(device) {}
 
 static ShapedBuffer AsShapedBuffer(
     int device_ordinal, const Shape& on_device_shape,
@@ -1119,22 +886,7 @@ PjRtFuture<Status> TfrtCpuBuffer::ToLiteral(MutableLiteralBase* literal) {
     return PjRtFuture<Status>(device_shape.status());
   }
   if (should_sync_copy) {
-    if (!on_device_shape().IsTuple()) {
-      const std::shared_ptr<MaybeOwningCpuMemory>& b =
-          device_buffer->Buffers()[0];
-      std::memcpy(literal->untyped_data(), b->data(),
-                  ShapeUtil::ByteSizeOf(*device_shape));
-    } else {
-      // Tuple case.
-      int num_leaves = literal->shape().tuple_shapes().size();
-      for (int i = 0; i < num_leaves; ++i) {
-        const std::shared_ptr<MaybeOwningCpuMemory>& b =
-            device_buffer->Buffers()[i];
-        std::memcpy(
-            literal->untyped_data({i}), b->data(),
-            ShapeUtil::ByteSizeOf(ShapeUtil::GetSubshape(*device_shape, {i})));
-      }
-    }
+    CopyCpuBufferToLiteral(*device_shape, device_buffer, literal);
     // Unblock ToLiteral caller.
     return PjRtFuture<Status>(OkStatus());
   } else {
@@ -1157,24 +909,7 @@ PjRtFuture<Status> TfrtCpuBuffer::ToLiteral(MutableLiteralBase* literal) {
               return;
             }
           }
-
-          if (!on_device_shape().IsTuple()) {
-            const std::shared_ptr<MaybeOwningCpuMemory>& b =
-                device_buffer->Buffers()[0];
-            std::memcpy(literal->untyped_data(), b->data(),
-                        ShapeUtil::ByteSizeOf(*device_shape));
-          } else {
-            // Tuple case.
-            int num_leaves = literal->shape().tuple_shapes().size();
-            for (int i = 0; i < num_leaves; ++i) {
-              const std::shared_ptr<MaybeOwningCpuMemory>& b =
-                  device_buffer->Buffers()[i];
-              std::memcpy(literal->untyped_data({i}), b->data(),
-                          ShapeUtil::ByteSizeOf(
-                              ShapeUtil::GetSubshape(*device_shape, {i})));
-            }
-          }
-
+          CopyCpuBufferToLiteral(*device_shape, device_buffer, literal);
           // Unblock ToLiteral event.
           ready_event.emplace(OkStatus());
         });
@@ -1287,57 +1022,6 @@ StatusOr<std::unique_ptr<PjRtBuffer>> TfrtCpuBuffer::CopyToDevice(
           on_device_shape_.IsTuple(), std::move(dst_buffers),
           std::move(dst_definition_events)),
       client(), tensorflow::down_cast<TfrtCpuDevice*>(dst_device)));
-}
-
-PjRtFuture<Status> TfrtCpuBuffer::GetReadyFuture() {
-  tfrt::AsyncValueRef<CpuEvent> definition_event;
-  {
-    absl::MutexLock lock(&mu_);
-    if (!tracked_device_buffer_) {
-      return PjRtFuture<Status>(InvalidArgument(
-          "GetReadyFuture() called on deleted or donated buffer"));
-    }
-    definition_event = tracked_device_buffer_->definition_event();
-  }
-  DCHECK(definition_event);
-
-  if (definition_event.IsAvailable()) {
-    if (definition_event.IsError()) {
-      return PjRtFuture<Status>(
-          FailedPrecondition("Buffer Definition Event: %s",
-                             definition_event.GetError().message()));
-    }
-    return PjRtFuture<Status>(OkStatus());
-  } else {
-    tfrt::AsyncValueRef<Status> status_event =
-        tfrt::MakeUnconstructedAsyncValueRef<Status>();
-
-    definition_event.AndThen(
-        [definition_event = definition_event.AsPtr(), status_event]() {
-          if (definition_event.IsError()) {
-            status_event.emplace(
-                FailedPrecondition("Buffer Definition Event: %s",
-                                   definition_event.GetError().message()));
-          } else {
-            status_event.emplace(OkStatus());
-          }
-        });
-
-    return PjRtFuture<Status>(
-        std::move(status_event),
-        /*on_block_start=*/
-        []() {
-          tsl::profiler::TraceMeProducer traceme("TfrtCpuBuffer::Await");
-          VLOG(1) << "TfrtCpuBuffer::Await";
-          return PjRtFutureHelpers::ProfilingKeys(
-              {/*traceme_context_id=*/traceme.GetContextId()});
-        },
-        /*on_block_end=*/
-        [](PjRtFutureHelpers::ProfilingKeys keys) {
-          tsl::profiler::TraceMeConsumer traceme("TfrtCpuBuffer::Await",
-                                                 keys.traceme_context_id);
-        });
-  }
 }
 
 TfrtCpuExecutable::TfrtCpuExecutable(
